Removed narrowing warnings by casting

+++ writer.cpp
@@ -13,7 +13,7 @@ namespace Writer {
         // Create WAV header
         WavHeader header = {
         {'R', 'I', 'F', 'F'},
-        sizeof(WavHeader) - 8 + fileSize,
+        static_cast<uint32_t>(sizeof(WavHeader)) - 8 + fileSize,
         {'W', 'A', 'V', 'E'},
         {'f', 'm', 't', ' '},
         16,
@@ -21,7 +21,7 @@ namespace Writer {
         num_channels,
         sample_rate,
         num_channels * sample_rate * bits_per_sample / 8,
-        num_channels * bits_per_sample / 8,
+        static_cast<uint16_t>(num_channels * bits_per_sample / 8),
         bits_per_sample,
         {'d', 'a', 't', 'a'},
         fileSize
